clean up change email
