+++ frontend/components/account/validations.js
@@ -1,4 +1,4 @@
-
+import firebase from '../../firebases';
 /**
  * signup validates all the input of the sign up form
  * @param {Object} props 
@@ -40,4 +40,41 @@ export const signin = props => {
   });
 
   return error_signin;
+}
+
+/**
+ * Change password validation
+ * @param {Object} props
+ */
+
+const password = ( password , newPassword) => {
+  let errors  = {};
+  console.log("password", password, 'new', newPassword);
+  if(!password) errors.password = "Please enter current password";
+  if(!password && password.length < 6) errors.password  = "Please enter a valid password!";
+  if(!newPassword) errors.newPassword = "Please enter a new password!";
+  if(!newPassword && newPassword.length < 6) errors.newPassword = "Please enter a valid password!"
+  if( password !== newPassword ) errors.newPassword = "Passwords DO NOT match";
+  console.log("errors", errors);
+  return errors;
+}
+
+/**
+ * Change email validation
+ * @param {Object} props
+ */
+
+const email = ( current, newEmail, password) => {
+  let errors  = {};
+  const userInfo = firebase.auth().currentUser;
+
+  const { email } = userInfo;
+  if(!current) errors.current = "Please enter current email!";
+  if(!current && current.length < 6) errors.current  = "Please enter current email!";
+  if(!newEmail) errors.newEmail = "Please enter a new email!";
+  if(!password) errors.password = "Please enter a password!";
+  if( !current && email !== current) errors.current = "This is not your current email!"
+
+  return errors;
+  
 }
